+++ src/main/kotlin/sk/v2/plugins/teamnotify/settings/TeamNotifySettingsFactory.kt
@@ -9,9 +9,3 @@ class TeamNotifySettingsFactory : ProjectSettingsFactory {
         return TeamNotifyProjectSettings()
     }
 }
-
-class DisabledWebhooksSettingsFactory : ProjectSettingsFactory {
-    override fun createProjectSettings(projectId: String): ProjectSettings {
-        return DisabledWebhooksSettings()
-    }
-}